requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --group ml -o requirements.txt
alembic==1.16.5
    # via mlflow
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via starlette
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via flask
boto3==1.40.35
    # via
    #   polyp-mlops (pyproject.toml)
    #   polyp-mlops (pyproject.toml:ml)
botocore==1.40.35
    # via
    #   boto3
    #   s3transfer
cachetools==5.5.2
    # via
    #   google-auth
    #   mlflow-skinny
certifi==2025.8.3
    # via
    #   kfp-server-api
    #   kubernetes
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.1.8
    # via
    #   click-option-group
    #   flask
    #   kfp
    #   mlflow-skinny
    #   ray
    #   uvicorn
click-option-group==0.5.7
    # via kfp
cloudpickle==3.1.1
    # via mlflow-skinny
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
databricks-sdk==0.65.0
    # via mlflow-skinny
docker==7.1.0
    # via mlflow
docstring-parser==0.17.0
    # via kfp
fastapi==0.117.1
    # via mlflow-skinny
filelock==3.19.1
    # via
    #   ray
    #   torch
flask==3.1.2
    # via mlflow
fonttools==4.60.0
    # via matplotlib
fsspec==2025.9.0
    # via torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via mlflow-skinny
google-api-core==2.25.1
    # via
    #   google-cloud-core
    #   google-cloud-storage
    #   kfp
google-auth==2.40.3
    # via
    #   databricks-sdk
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
    #   kfp
    #   kubernetes
google-cloud-core==2.4.3
    # via google-cloud-storage
google-cloud-storage==3.4.0
    # via kfp
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.70.0
    # via google-api-core
graphene==3.4.3
    # via mlflow
graphql-core==3.2.6
    # via
    #   graphene
    #   graphql-relay
graphql-relay==3.2.0
    # via graphene
greenlet==3.2.4
    # via sqlalchemy
grpcio==1.75.0
    # via ray
gunicorn==23.0.0
    # via mlflow
h11==0.16.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
importlib-metadata==8.7.0
    # via
    #   mlflow-skinny
    #   opentelemetry-api
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   flask
    #   torch
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
joblib==1.5.2
    # via scikit-learn
jsonschema==4.25.1
    # via ray
jsonschema-specifications==2025.9.1
    # via jsonschema
kfp==2.14.3
    # via polyp-mlops (pyproject.toml:ml)
kfp-pipeline-spec==2.14.0
    # via kfp
kfp-server-api==2.14.3
    # via kfp
kiwisolver==1.4.9
    # via matplotlib
kubernetes==30.1.0
    # via kfp
loguru==0.7.3
    # via polyp-mlops (pyproject.toml:ml)
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   mako
    #   werkzeug
matplotlib==3.10.6
    # via
    #   polyp-mlops (pyproject.toml:ml)
    #   mlflow
mlflow==3.1.0
    # via polyp-mlops (pyproject.toml:ml)
mlflow-skinny==3.1.0
    # via mlflow
mpmath==1.3.0
    # via sympy
msgpack==1.1.1
    # via ray
networkx==3.5
    # via torch
numpy==2.1.2
    # via
    #   polyp-mlops (pyproject.toml:ml)
    #   contourpy
    #   matplotlib
    #   mlflow
    #   pandas
    #   scikit-learn
    #   scipy
    #   torchvision
nvidia-cublas-cu12==12.6.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.6.80
    # via torch
nvidia-cuda-nvrtc-cu12==12.6.77
    # via torch
nvidia-cuda-runtime-cu12==12.6.77
    # via torch
nvidia-cudnn-cu12==9.5.1.17
    # via torch
nvidia-cufft-cu12==11.3.0.4
    # via torch
nvidia-cufile-cu12==1.11.1.6
    # via torch
nvidia-curand-cu12==10.3.7.77
    # via torch
nvidia-cusolver-cu12==11.7.1.2
    # via torch
nvidia-cusparse-cu12==12.5.4.2
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.3
    # via torch
nvidia-nccl-cu12==2.26.2
    # via torch
nvidia-nvjitlink-cu12==12.6.85
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.6.77
    # via torch
oauthlib==3.3.1
    # via
    #   kubernetes
    #   requests-oauthlib
opentelemetry-api==1.37.0
    # via
    #   mlflow-skinny
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-sdk==1.37.0
    # via mlflow-skinny
opentelemetry-semantic-conventions==0.58b0
    # via opentelemetry-sdk
packaging==25.0
    # via
    #   gunicorn
    #   matplotlib
    #   mlflow-skinny
    #   ray
pandas==2.3.2
    # via mlflow
pillow==11.3.0
    # via
    #   matplotlib
    #   torchvision
proto-plus==1.26.1
    # via google-api-core
protobuf==6.31.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   kfp
    #   kfp-pipeline-spec
    #   mlflow-skinny
    #   proto-plus
    #   ray
pyarrow==20.0.0
    # via
    #   polyp-mlops (pyproject.toml:ml)
    #   mlflow
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.9
    # via
    #   fastapi
    #   mlflow-skinny
pydantic-core==2.33.2
    # via pydantic
pyparsing==3.2.5
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   graphene
    #   kfp-server-api
    #   kubernetes
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via polyp-mlops (pyproject.toml:ml)
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   kfp
    #   kubernetes
    #   mlflow-skinny
    #   ray
ray==2.48.0
    # via polyp-mlops (pyproject.toml:ml)
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   databricks-sdk
    #   docker
    #   google-api-core
    #   google-cloud-storage
    #   kubernetes
    #   mlflow-skinny
    #   ray
    #   requests-oauthlib
    #   requests-toolbelt
requests-oauthlib==2.0.0
    # via kubernetes
requests-toolbelt==1.0.0
    # via kfp
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
s3transfer==0.14.0
    # via boto3
scikit-learn==1.7.2
    # via mlflow
scipy==1.16.2
    # via
    #   mlflow
    #   scikit-learn
setuptools==80.9.0
    # via
    #   torch
    #   triton
six==1.17.0
    # via
    #   kfp-server-api
    #   kubernetes
    #   python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via
    #   alembic
    #   mlflow
sqlparse==0.5.3
    # via mlflow-skinny
starlette==0.48.0
    # via fastapi
sympy==1.14.0
    # via torch
tabulate==0.9.0
    # via kfp
thop==0.1.1.post2209072238
    # via polyp-mlops (pyproject.toml:ml)
threadpoolctl==3.6.0
    # via scikit-learn
torch==2.7.1
    # via
    #   polyp-mlops (pyproject.toml:ml)
    #   thop
    #   torchvision
torchvision==0.22.1
    # via polyp-mlops (pyproject.toml:ml)
tqdm==4.67.1
    # via
    #   polyp-mlops (pyproject.toml)
    #   polyp-mlops (pyproject.toml:ml)
triton==3.3.1
    # via torch
typing-extensions==4.15.0
    # via
    #   alembic
    #   fastapi
    #   graphene
    #   grpcio
    #   mlflow-skinny
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   botocore
    #   docker
    #   kfp
    #   kfp-server-api
    #   kubernetes
    #   requests
uvicorn==0.36.0
    # via mlflow-skinny
websocket-client==1.8.0
    # via kubernetes
werkzeug==3.1.3
    # via flask
zipp==3.23.0
    # via importlib-metadata
